Solution to Calculate and Print the Area of a Circle

+++ Calculate and Print the Area of a Circle@@6788d5012db25efec946f943/code.c
@@ -1,10 +1,10 @@
 #include <stdio.h>
 
 int main() {
-    int radius;
+    float radius;
     float area;
 
-    scanf("%f", &radius);
+    scanf("%d", &radius);
 
     area = 3.14 * radius * radius;
